refactor: 미사용 코드 삭제 및 localStorage 데이터 상수로 변경
