+++ src/app/(route)/mypage/_components/IdolListSwiper.tsx
