+++ src/app/(route)/mypage/_constants/mypage.constants.ts
@@ -1,1 +1,7 @@
+import type { IdolData } from '@/types/idols.interface';
+
 export const LOCAL_STORAGE_KEY = 'favoriteIdol';
+
+export const LOCAL_STORAGE_DATA: IdolData[] = JSON.parse(
+  localStorage?.getItem(LOCAL_STORAGE_KEY) ?? '[]',
+);
